+++ test/loader/test_link_neighbor_loader.py
@@ -3,7 +3,6 @@ import torch
 
 from torch_geometric.data import Data, HeteroData
 from torch_geometric.loader import LinkNeighborLoader
-from torch_geometric.testing import withPackage
 from torch_geometric.testing.feature_store import MyFeatureStore
 from torch_geometric.testing.graph_store import MyGraphStore
 
@@ -182,7 +181,6 @@ def test_link_neighbor_loader_edge_label():
         assert torch.all(batch.edge_label[10:] == 0)
 
 
-@withPackage('torch_sparse>=0.6.14')
 def test_temporal_heterogeneous_link_neighbor_loader():
     data = HeteroData()